# Compilation speedup of traversal