--- a/src/traversal_encoding.jl
+++ b/src/traversal_encoding.jl
@@ -61,10 +61,10 @@ _pred_traversal(::LeafNode, n, p, s="") = p(n) ? [stringify(s)] : String[]
 function _pred_traversal(::InnerNode, n, p, s="")
     d = printchildren(n)
     l = length(d)
-    z = Vector{Vector{String}}[pred_traversal(_ith_child(d, i), p, s * encode(i, l)) for i in 1:l]
-    res = vcat(z...)
+    z = Vector{String}[pred_traversal(_ith_child(d, i), p, s * encode(i, l)) for i in 1:l]
+    res = isempty(z) ? String[] : reduce(vcat, z)
     p(n) ? vcat(stringify(s), res) : res
-end 
+end
 
 list_traversal(n) = pred_traversal(n, t -> true)
 find_traversal(n, x) = pred_traversal(n, t -> x === t)